requirements.txt
streamlit
tweepy
vaderSentiment
python-dotenv